product blueprint added
