+++ flaskr/templates/layout.html
@@ -5,7 +5,7 @@
 </head>
 <body>
     {% block body %}
-        <a href=/home>Home</a><br/>
+        <a href=/>Home</a><br/>
         <a href=/about>About Us</a><br/>
         <a href=/products>Products</a><br/>
         <a href=/create_product>Create Product</a><br/>
